requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
pycryptodome==3.10.1
Werkzeug==2.0.3
